requirements.txt
flet==0.25.2
python-dotenv==1.0.1
fastapi
uvicorn
